api done, falta testeo

+++ app/Http/Controllers/StockController.php
@@ -15,15 +15,15 @@ class StockController extends Controller
      *
      * @return JsonResponse
      */
-    public function index(): JsonResponse
+    public function index(Request $request): JsonResponse
     {
-        $stock = Stock::all();
-        return response()->json([
+
+        $sal_id = $request->get('sal_id');
+        $stock = Stock::where('sal_id', $sal_id)->get();       
+         return response()->json([
             'status' => 200,
             'message' => "Exitoso",
-            'data' => [
-                'stock' => $stock,
-            ]
+            'stock' => $stock,
         ]);
     }
 
@@ -35,30 +35,27 @@ class StockController extends Controller
      */
     public function create(Request $request)
     {
+
         try {
             DB::beginTransaction();
-
-
-            $stock = new Stock();
-            $stock = $stock->create($request->all());
+                $sal_id = $request->get('sal_id');
+                $stock = new Stock;
+                $stock->sal_id = $sal_id;
+                $stock = $stock->create(array_merge($request->all(), ['sal_id' => $sal_id]));
 
 
             DB::commit();
             return response()->json([
                 'status' => 200,
                 'message' => "Exitoso",
-                'data' => [
-                    'stock' => $stock,
-                ]
+                'stock' => $stock,
             ]);
         } catch (Exception $e) {
             DB::rollBack();
             return response()->json([
                 'status' => 500,
                 'message' => "Error at creating stock",
-                'data' => [
-                    'error' => $e->getMessage()
-                ]
+                'error' => $e->getMessage()
             ]);
         }
 
@@ -70,35 +67,38 @@ class StockController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  Stock  $stock
+     * @param  $sto_id
      * @return JsonResponse
      */
-    public function update(Request $request, $ser_id)
+    public function update(Request $request, $sto_id)
     {
         try {
             DB::beginTransaction();
-            $stock = Stock::find($ser_id);
-            $stock->update($request->all());
-            DB::commit();
-            return response()->json([
-                'status' => 200,
-                'message' => "Exitoso",
-                'data' => [
-                    'stock' => $stock,
-                ]
-            ]);
+            $sal_id = $request->get('sal_id');
+            if(Stock::where('sal_id', $sal_id)->where('sto_id', $sto_id)->exists()){
+             $stock = Stock::where('sal_id', $sal_id)->where('sto_id', $sto_id)->first();
+             $stock->update($request->all());
+             DB::commit();
+                 return response()->json([
+                        'status' => 200,
+                        'message' => "Exitoso",
+                        'stock' => $stock,
+                 ]);
+            }else{
+                 return response()->json([
+                        'status' => 400,
+                        'message' => "Error, no se encuentra el producto",
+                 ],400);
+            }
+
         }catch (Exception $e){
             DB::rollBack();
             return response()->json([
                 'status' => 500,
                 'message' => "Error",
-                'data' => [
-                    'error' => $e->getMessage(),
-                ]
-            ]);
+                'error' => $e->getMessage(),
+            ],500);
         }
-
-
     }
     /**
      * Display a listing of the resource.
@@ -131,33 +131,57 @@ class StockController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param $ser_id
+     * @param $sto_id
      * @return JsonResponse
      */
-    public function destroy($ser_id)
+    public function destroy($sto_id)
     {
         try {
-            DB::beginTransaction();
-            $stock = Stock::find($ser_id);
-            $stock->delete();
-            DB::commit();
-            return response()->json([
-                'status' => 200,
-                'message' => "Exitoso item delete",
-                'data'=>[
-                    'item'=>$stock
-                ]
+             DB::beginTransaction();
+              $sal_id = $request->get('sal_id');
+ 
+              if(Stock::where('sal_id', $sal_id)->where('sto_id', $sto_id)->exists()){
+                 $stock = Stock::where('sal_id', $sal_id)->where('sto_id', $sto_id)->first();
+                 $stock->delete();
+             }else{
+                 return response()->json([
+                       'status' => 400,
+                       'message' => "Error, no se encuentra el producto",
+                       ]);
+             }
+ 
+             DB::commit();
+             return response()->json([
+                 'status' => 200,
+                 'message' => "Exitoso stock delete",
+             ]);
+         }catch (Exception $e){
+             DB::rollBack();
+             return response()->json([
+                 'status' => 500,
+                 'message' => "Error",
+                 'error' => $e->getMessage(),
+             ]);
+         }
+     }
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function listStockByServicesByAppointment(Request $request, $sto_id): JsonResponse
+    {
+        $stock = DB::table('stocks')
+            ->join('stock__by__services__appointments','stocks.sto_id','=','stock__by__services__appointments.sto_id')
+            ->select('stocks.*')
+            ->where('stocks.sto_id',$sto_id)
+            ->get();
+        return response()->json([
+            'status' => 200,
+            'message' => "Exitoso",
+            'stock' => $stock,
 
-            ]);
-        }catch (Exception $e){
-            DB::rollBack();
-            return response()->json([
-                'status' => 500,
-                'message' => "Error",
-                'data' => [
-                    'error' => $e->getMessage(),
-                ]
-            ]);
-        }
+        ]);
     }
+
 }
